+++ NEWS.md
@@ -1,5 +1,8 @@
 # rnpn (development version)
 
+* `rnpn` depends on the `terra` package instead of `raster`
+* `nnpn` no longer depends on the `sp` package
+
 # rnpn 1.2.9 (2024-08-18)
 
 ### NEW FEATURES